+++ L21_SpringBoot_adv/src/main/java/ru/selezneva/L21_SpringBoot_adv/atm/exceptions/AmountCannotBeCollected.java
@@ -1,4 +1,4 @@
 package ru.selezneva.L21_SpringBoot_adv.atm.exceptions;
 
-public class AmountCannotBeCollected extends RuntimeException {
+public class AmountCannotBeCollected extends ATMException {
 }
