+++ L21_SpringBoot_adv/src/main/java/ru/selezneva/L21_SpringBoot_adv/atm/exceptions/DownloadATMExeption.java
@@ -1,6 +1,6 @@
 package ru.selezneva.L21_SpringBoot_adv.atm.exceptions;
 
-public class DownloadATMExeption extends RuntimeException {
+public class DownloadATMExeption extends ATMException {
     public DownloadATMExeption(Throwable cause) {
         super(cause);
     }
