+++ L21_SpringBoot_adv/src/main/java/ru/selezneva/L21_SpringBoot_adv/atm/exceptions/SavingATMToFileException.java
@@ -1,6 +1,6 @@
 package ru.selezneva.L21_SpringBoot_adv.atm.exceptions;
 
-public class SavingATMToFileException extends RuntimeException {
+public class SavingATMToFileException extends ATMException {
     public SavingATMToFileException(Throwable cause) {
         super(cause);
     }
